Improves reset functionality and repo initialization
Updates the reset functionality to more robustly remove
Also this code will literally kill my local repo if I run it so I'm not planning to do that in this moment
contents and prevents deletion of java files. Ensures that
the repository is initialized after cleaning up the
testing environment. This prevents failures by re-initializing
the repository as needed.