GP-4.2
I changed Createtreefromindex so it now returns the roothash as a string. I added a commit method that puts a commit file into the objects folder for each time the user commits. inside the commit file there is the author, the tree, a message, a previous commit, and the date and time. After the commit gets added it adds itself to the head file so it becomes the most recent commit